backend/auth/pyproject.toml
[project]
name = "fastapi-auth"
version = "0.1.0"
description = ""
authors = [{ name = "nikolaev38", email = "kulakovsky@at-signal.ru" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "loki-logger-handler (>=1.1.2,<2.0.0)",
    "pytz (>=2025.2,<2026.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
